requirements.txt
Django==3.1.1
django-environ==0.4.5
sorl-thumbnail==12.6.3
Pillow==7.2.0
djangorestframework==3.11.1
asgiref==3.2.10
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
django-debug-toolbar==2.2
flake8
black
pytest-django==3.10.0